Filter changes on Post/fact-check page
