+++ src/components/List/index.js
@@ -1,157 +1,34 @@
 import React, { useState } from 'react';
-import { Popconfirm, Button, List, Input, Select, Form, Space, Tag, Row, Col } from 'antd';
+import { Popconfirm, Button, List, Space, Tag } from 'antd';
 import { EditOutlined, DeleteOutlined } from '@ant-design/icons';
-import { useDispatch, useSelector } from 'react-redux';
-import { getPosts, deletePost } from '../../actions/posts';
+import { useDispatch } from 'react-redux';
+import { deletePost } from '../../actions/posts';
 import { Link } from 'react-router-dom';
-import Selector from '../../components/Selector';
-import deepEqual from 'deep-equal';
-import Template from '../../components/Template';
 import ImagePlaceholder from '../../components/ErrorsAndImage/PlaceholderImage';
 import QuickEdit from './QuickEdit';
 import moment from 'moment';
 
-function PostList({ actions, format }) {
+function PostList({ actions, format, filters, setFilters, data, fetchPosts }) {
   const dispatch = useDispatch();
-  const { Option } = Select;
-  const [form] = Form.useForm();
-
-  const [filters, setFilters] = React.useState({
-    page: 1,
-    limit: 20,
-    format: [format.id],
-  });
   const [id, setID] = useState(0);
 
-  const { posts, total, loading, tags, categories } = useSelector((state) => {
-    const node = state.posts.req.find((item) => {
-      return deepEqual(item.query, filters);
-    });
-
-    if (node)
-      return {
-        posts: node.data.map((element) => {
-          const post = state.posts.details[element];
-
-          post.medium = state.media.details[post.featured_medium_id];
-          return post;
-        }),
-        total: node.total,
-        loading: state.posts.loading,
-        tags: state.tags.details,
-        categories: state.categories.details,
-      };
-    return { posts: [], total: 0, loading: state.posts.loading, tags: {}, categories: {} };
-  });
-
-  React.useEffect(() => {
-    fetchPosts();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [filters]);
-
-  const fetchPosts = () => {
-    dispatch(getPosts(filters));
-  };
-
-  const onSave = (values) => {
-    let filterValue = {
-      tag: values.tags,
-      category: values.categories,
-      sort: values.sort,
-      q: values.q,
-      author: values.authors,
-      status: values.status !== 'all' ? values.status : null,
-    };
-
-    setFilters({ ...filters, ...filterValue });
-  };
-
   const getTagList = (tagids) => {
-    return tagids.map((id) => <Tag>{tags[id].name}</Tag>);
+    return tagids.map((id) => <Tag>{data.tags[id].name}</Tag>);
   };
   const getCategoryList = (catIds) => {
-    return catIds.map((id) => <Tag>{categories[id].name}</Tag>);
+    return catIds.map((id) => <Tag>{data.categories[id].name}</Tag>);
   };
 
   return (
     <Space direction="vertical">
-      <Template format={format} />
-      <Form
-        initialValues={filters}
-        form={form}
-        name="filters"
-        onFinish={(values) => onSave(values)}
-        style={{ maxWidth: '100%' }}
-        className="ant-advanced-search-form"
-        onValuesChange={(changedValues, allValues) => {
-          console.log('changedValues', changedValues, 'all', allValues);
-          if (!changedValues.q) {
-            onSave(allValues);
-          }
-        }}
-      >
-        <Row gutter={10}>
-          <Col span={4} key={1} offset={12}>
-            <Form.Item name="q">
-              <Input placeholder="search posts" />
-            </Form.Item>
-          </Col>
-
-          <Col>
-            <Form.Item>
-              <Button htmlType="submit">Search</Button>
-            </Form.Item>
-          </Col>
-          <Col key={2}>
-            <Form.Item name="sort" label="Sort">
-              <Select defaultValue="desc" style={{ width: '100%' }}>
-                <Option value="desc">Latest</Option>
-                <Option value="asc">Old</Option>
-              </Select>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={10}>
-          <Col span={5} key={3}>
-            <Form.Item name="status" label="Status">
-              <Select defaultValue="all">
-                <Option value="all">All</Option>
-                <Option value="draft">Draft</Option>
-                <Option value="publish">Publish</Option>
-                <Option value="ready">Ready to Publish</Option>
-              </Select>
-            </Form.Item>
-          </Col>
-          <Col span={5} key={4}>
-            <Form.Item name="tags" label="Tags">
-              <Selector mode="multiple" action="Tags" placeholder="Filter Tags" />
-            </Form.Item>
-          </Col>
-          <Col span={5} key={5}>
-            <Form.Item name="categories" label="Categories">
-              <Selector mode="multiple" action="Categories" placeholder="Filter Categories" />
-            </Form.Item>
-          </Col>
-          <Col span={5} key={6}>
-            <Form.Item name="authors" label="Authors">
-              <Selector
-                mode="multiple"
-                action="Authors"
-                placeholder="Filter Authors"
-                display={'email'}
-              />
-            </Form.Item>
-          </Col>
-        </Row>
-      </Form>
       <List
         bordered
         className="post-list"
-        loading={loading}
+        loading={data.loading}
         itemLayout="vertical"
-        dataSource={posts}
+        dataSource={data.posts}
         pagination={{
-          total: total,
+          total: data.total,
           current: filters.page,
           pageSize: filters.limit,
           onChange: (pageNumber, pageSize) =>
